refactor: main page
- QuizBanner 내부 컴포넌트들 별도 파일로 분리 (코드 가독성을 위해)
